Move Jetpack logo to icon file
Forward compat with more icons to come. See #15717.
